pak configuration

Environment variables and options that modify the default pak behavior.

Environment variables

PKG_BUILD_VIGNETTES

TODO

PKG_CACHE_DIR

TODO

PKG_CRAN_MIRROR

TODO

PKG_DEPENDENCIES

TODO

PKG_LIBRARY

TODO

PKG_METADATA_CACHE_DIR

TODO

PKG_METADATA_UPDATE_AFTER

TODO

PKG_PLATFORMS

TODO

PKG_R_VERSIONS

TODO

PKG_SYSREQS

TODO

PKG_SYSREQS_RSPM_REPO_ID

TODO

PKG_SYSREQS_RSPM_URL

TODO

PKG_SYSREQS_SUDO

TODO

PKG_SYSREQS_VERBOSE

TODO

PKG_WINDOWS_ARCHS

TODO

Options

Ncpus

Set to the desired number of worker processes for package installation. If not set, then pak will use the number of logical processors in the machine.

pak.no_extra_messages

Set to TRUE to omit pak's warnings about missing extra packages.

pkg.build_vignettes

TODO

pkg.cache_dir

TODO

pkg.cran_mirror

TODO

pkg.dependencies

TODO

pkg.library

TODO

pkg.metadata_cache_dir

TODO

pkg.metadata_update_after

TODO

pkg.package_cache_dir

TODO

pkg.platforms

TODO

pkg.r_versions

TODO

pkg.sysreqs

TODO

pkg.sysreqs_rspm_repo_id

TODO

pkg.sysreqs_rspm_url

TODO

pkg.sysreqs_sudo

TODO

pkg.sysreqs_verbose

TODO

pkg.windows_archs

TODO

repos

The CRAN-like repositories to use. See ?options for details.