local_package_trees {pak}	R Documentation

About local package trees

Description

pak can install packages from local package trees. This is convenient for package development. See the following functions:

local_install() installs a package from a package tree and all of its dependencies.

local_install_deps() installs all hard dependencies of a package.

local_install_dev_deps() installs all hard and soft dependencies of a package. This function is intended for package development.

Details

Note that the last two functions do not install the package in the specified package tree itself, only its dependencies. This is convenient if the package itself is loaded via some other means, e.g. devtools::load_all(), for development.

See Also

Other local package trees: local_deps_explain(), local_deps(), local_install_deps(), local_install_dev_deps(), local_install(), pak()

[Package pak version 0.9.2 Index]