pak {pak}	R Documentation

Install specified required packages

Description

Install the specified packages, or the ones required by the package or project in the current working directory.

Usage

pak(pkg = NULL, ...)

Arguments

pkg
Package names or remote package specifications to install. See pak package sources for details. If NULL, will install all development dependencies for the current package.

...
Extra arguments are passed to pkg_install() or local_install_dev_deps().

Details

This is a convenience function:

If you want to install some packages, it is easier to type than pkg_install().

If you want to install all the packages that are needed for the development of a package or project, then it is easier to type than local_install_dev_deps().

You don't need to remember two functions to install packages, just one.

See Also

Other package functions: lib_status(), pkg_deps_tree(), pkg_deps(), pkg_download(), pkg_install(), pkg_remove(), pkg_status(), pkg_sysreqs()

Other local package trees: local_deps_explain(), local_deps(), local_install_deps(), local_install_dev_deps(), local_install(), local_package_trees

[Package pak version 0.9.2 Index]