pkg_sysreqs {pak}	R Documentation

Calculate system requirements of one of more packages

Description

Calculate system requirements of one of more packages

Usage

pkg_sysreqs(pkg, upgrade = TRUE, dependencies = NA, sysreqs_platform = NULL)

Arguments

pkg
Package names or package references. E.g.

ggplot2: package from CRAN, Bioconductor or a CRAN-like repository in general,

tidyverse/ggplot2: package from GitHub,

tidyverse/ggplot2@v3.4.0: package from GitHub tag or branch,

https://examples.com/.../ggplot2_3.3.6.tar.gz: package from URL,

.: package in the current working directory.

See "Package sources" for more details.

upgrade
When FALSE, the default, pak does the minimum amount of work to give you the latest version(s) of pkg. It will only upgrade dependent packages if pkg, or one of their dependencies explicitly require a higher version than what you currently have. It will also prefer a binary package over to source package, even it the binary package is older.

When upgrade = TRUE, pak will ensure that you have the latest version(s) of pkg and all their dependencies.

dependencies
What kinds of dependencies to install. Most commonly one of the following values:

NA: only required (hard) dependencies,

TRUE: required dependencies plus optional and development dependencies,

FALSE: do not install any dependencies. (You might end up with a non-working package, and/or the installation might fail.) See Package dependency types for other possible values and more information about package dependencies.

sysreqs_platform
System requirements platform.

If NULL, then the sysreqs_platform configuration option is used, which defaults to the current platform.

Set this option if to one of "alpine", "centos", "debian", "fedora", "opensuse", "opensuse", "opensuse", "redhat", "redhat", "redhat", "redhat", "redhat", "redhat", "rockylinux", "rockylinux", "rockylinux", "sle", "sle", "ubuntu" if pak fails to correctly detect your platform or if you want to see the system requirements for a different platform.

Value

List with entries:

os: character string. Operating system.

distribution: character string. Linux distribution, NA if the OS is not Linux.

version: character string. Distribution version, NA is the OS is not Linux.

pre_install: character vector. Commands to run before the installation of system packages.

install_scripts: character vector. Commands to run to install the system packages.

post_install: character vector. Commands to run after the installation of system packages.

packages: data frame. Information about the system packages that are needed. It has columns:

sysreq: string, cross-platform name of the system requirement.

packages: list column of character vectors. The names of the R packages that have this system requirement.

pre_install: list column of character vectors. Commands run before the package installation for this system requirement.

system_packages: list column of character vectors. Names of system packages to install.

post_install: list column of character vectors. Commands run after the package installation for this system requirement.

See Also

Other package functions: lib_status(), pak(), pkg_deps(), pkg_deps_tree(), pkg_download(), pkg_install(), pkg_remove(), pkg_status()

Other system requirements functions: sysreqs_check_installed(), sysreqs_db_list(), sysreqs_db_match(), sysreqs_db_update(), sysreqs_is_supported(), sysreqs_list_system_packages(), sysreqs_platforms()

[Package pak version 0.9.2 Index]