ppm_platforms {pak}	R Documentation

List all platforms supported by Posit Package Manager (PPM)

Description

List all platforms supported by Posit Package Manager (PPM)

Usage

ppm_platforms()

Value

Data frame with columns:

name: platform name, this is essentially an identifier,

os: operating system, linux, windows or macOS currently,

binary_url: the URL segment of the binary repository URL of this platform, see ppm_snapshots().

distribution: for Linux platforms the name of the distribution,

release: for Linux platforms, the name of the release,

binaries: whether PPM builds binaries for this platform.

See Also

The 'pkgcache and Posit Package Manager on Linux' article at https://r-lib.github.io/pkgcache/.

Other PPM functions: ppm_has_binaries(), ppm_r_versions(), ppm_repo_url(), ppm_snapshots()

Examples

ppm_platforms()

[Package pak version 0.9.2 Index]