Install project dependencies into the project library

The project library is in r-packages, within the project directory.

proj_install(pkg = NULL, root = ".", upgrade = FALSE,
  optional = FALSE, ask = interactive())

Arguments

pkg
Package(s) to install. By default (if NULL), they are taken from the DESCRIPTION file in root. If not NULL, then the packages will be added to the DESCRIPTION files, and existing references to these packages will be removed first.

root
Path to the project directory. A subdirectory can be given as well. Defaults to the current working directory.

upgrade
Whether to try to install the latest available versions of the specified package(s) and/or all dependencies.

optional
If pkg is not NULL, then the new packages will be added as optional, i.e. in the Suggests section.

ask
Whether to ask the user for confirmation. For non-interactive installs supplyFALSE here.

Value

Data frame containing data about the installed / updated packages.

See also