sysreqs_check_installed {pak}	R Documentation

Check if installed packages have all their system requirements

Description

sysreqs_check_installed() checks if the system requirements of all packages (or a subset of packages) are installed.

sysreqs_fix_installed() installs the missing system packages.

Usage

sysreqs_check_installed(packages = NULL, library = .libPaths()[1])
sysreqs_fix_installed(packages = NULL, library = .libPaths()[1])

Arguments

packages
If not NULL, then only these packages are checked. If a package in packages is not installed, then pak throws a warning.

library
Library or libraries to check.

Details

These functions use the sysreqs_platform configuration option, see Configuration. Set this if pak does not detect your platform correctly.

Value

Data frame with a custom print and format method, and a pkg_sysreqs_check_result class. Its columns are:

system_package: string, name of the required system package.

installed: logical, whether the system package is correctly installed.

packages: list column of character vectors. The names of the installed R packages that need this system package.

pre_install: list column of character vectors. Commands to run before the installation of the the system package.

post_install: list column of character vectors. Commands to run after the installation of the system package.

The data frame also have two attributes with additional data:

sysreqs_records: the raw system requirements records, and

system_packages: the list of the installed system packages.

sysreqs_fix_packages() returns the same value, but invisibly.

See Also

Other system requirements functions: pkg_sysreqs(), sysreqs_db_list(), sysreqs_db_match(), sysreqs_db_update(), sysreqs_is_supported(), sysreqs_list_system_packages(), sysreqs_platforms()

Examples

# This only works on supported platforms
sysreqs_check_installed()

[Package pak version 0.9.2 Index]