sysreqs_db_list {pak}	R Documentation

List contents of the system requirements DB, for a platform

Description

It also tries to update the system dependency database, if it is outdated. (I.e. older than allowed in the metadata_update_after configuration option.

Usage

sysreqs_db_list(sysreqs_platform = NULL)

Arguments

sysreqs_platform
System requirements platform. If NULL, then the sysreqs_platform configuration option is used, which defaults to the current platform. Set this option if pak does not detect your platform correctly.

Value

Data frame with columns:

name: cross platform system dependency name in the database.

patterns: one or more regular expressions to match to SystemRequirements fields.

packages: one or more system package names to install.

pre_install: command(s) to run before installing the packages.

post_install:: command(s) to run after installing the packages.

See Also

Other system requirements functions: pkg_sysreqs(), sysreqs_check_installed(), sysreqs_db_match(), sysreqs_db_update(), sysreqs_is_supported(), sysreqs_list_system_packages(), sysreqs_platforms()

Examples

sysreqs_db_list(sysreqs_platform = "ubuntu-22.04")

[Package pak version 0.9.2 Index]