sysreqs_db_match {pak}	R Documentation

Match system requirement descriptions to the database

Description

In the usual workflow pak matches the SystemRequirements fields of the DESCRIPTION files to the database.

Usage

sysreqs_db_match(specs, sysreqs_platform = NULL)

Arguments

specs
Character vector of system requirements descriptions.

sysreqs_platform
System requirements platform. If NULL, then the sysreqs_platform configuration option is used, which defaults to the current platform. Set this option if pak does not detect your platform correctly.

Details

The sysreqs_db_match() function lets you match any string, and it is mainly useful for debugging.

Value

Data frame with columns:

spec: the input specs.

sysreq: name of the system library or tool.

packages: system packages, list column of character vectors. Rarely it can be an empty string, e.g. if a pre_install script performs the installation.

pre_install: list column of character vectors. Shell script(s) to run before the installation.

post_install: list column of character vectors. Shell script(s) to run after the installation.

See Also

Other system requirements functions: pkg_sysreqs(), sysreqs_check_installed(), sysreqs_db_list(), sysreqs_db_update(), sysreqs_is_supported(), sysreqs_list_system_packages(), sysreqs_platforms()

Examples

sysreqs_db_match(
  c("Needs libcurl", "Java, libssl"),
  sysreqs_platform = "ubuntu-22.04"
)

[Package pak version 0.9.2 Index]