sysreqs_db_update {pak}	R Documentation

Update the cached copy of the system requirements database

Description

Update the cached copy of the system requirements database

Usage

sysreqs_db_update()

Details

If the the cached copy is recent, then no update is attempted. See the metadata_update_after configuration option.

See Also

Other system requirements functions: pkg_sysreqs(), sysreqs_check_installed(), sysreqs_db_list(), sysreqs_db_match(), sysreqs_is_supported(), sysreqs_list_system_packages(), sysreqs_platforms()

[Package pak version 0.9.2 Index]