sysreqs_is_supported {pak}	R Documentation

Check if a platform has system requirements support

Description

Check if a platform has system requirements support

Usage

sysreqs_is_supported(sysreqs_platform = NULL)

Arguments

sysreqs_platform
System requirements platform. If NULL, then the sysreqs_platform configuration option is used, which defaults to the current platform. Set this option if pak does not detect your platform correctly.

Value

Logical scalar.

See Also

The sysreqs_platform configuration option.

Other system requirements functions: pkg_sysreqs(), sysreqs_check_installed(), sysreqs_db_list(), sysreqs_db_match(), sysreqs_db_update(), sysreqs_list_system_packages(), sysreqs_platforms()

Examples

sysreqs_is_supported()

[Package pak version 0.9.2 Index]