sysreqs_list_system_packages {pak}	R Documentation

List installed system packages

Description

List installed system packages

Usage

sysreqs_list_system_packages()

Details

This function uses the sysreqs_platform configuration option, see Configuration. Set this if pak does not detect your platform correctly.

Value

Data frame with columns:

status. two or three characters, the notation of dpkg on Debian based systems. "ii" means the package is correctly installed. On RPM based systems it is always "ii" currently.

package: name of the system package.

version: installed version of the system package.

capabilities: list column of character vectors, the capabilities provided by the package.

See Also

Other system requirements functions: pkg_sysreqs(), sysreqs_check_installed(), sysreqs_db_list(), sysreqs_db_match(), sysreqs_db_update(), sysreqs_is_supported(), sysreqs_platforms()

Examples

sysreqs_list_system_packages()[1:10,]

[Package pak version 0.9.2 Index]