sysreqs_platforms {pak}	R Documentation

List platforms with system requirements support

Description

List platforms with system requirements support

Usage

sysreqs_platforms()

Value

Data frame with columns:

name: human readable OS name.

os: OS name, e.g. linux.

distribution: OS id, e.g. ubuntu or redhat.

version: distribution version. A star means that all versions are supported, that are also supported by the vendor.

update_command: command to run to update the system package metadata.

install_command: command to run to install packages.

query_command: name of the tool to use to query system package information.

See Also

Other system requirements functions: pkg_sysreqs(), sysreqs_check_installed(), sysreqs_db_list(), sysreqs_db_match(), sysreqs_db_update(), sysreqs_is_supported(), sysreqs_list_system_packages()

Examples

sysreqs_platforms()

[Package pak version 0.9.2 Index]