Install the dependencies of a package tree

Installs the hard dependencies of a package tree (or source package file), without installing the package tree itself.

local_install_deps(root = ".", lib = .libPaths()[1], upgrade = FALSE,
  ask = interactive())

Arguments

root
Path to the package tree.

lib
Package library to install the packages to. Note that all dependent packages will the be installed here, even if they are already installed in another library.

upgrade
Whether to upgrade already installed packages to the latest available version. If this is FALSE, then only packages that need updates to satisfy version requirements, will be updated. If it is TRUE, all specified or dependent packages will be updated to the latest available version.

ask
Whether to ask for confirmation.

Value

Data frame, with information about the installed package(s).

Details

Note that development (and optional) dependencies, under Suggests in DESCRIPTION, are not installed. If you want to install them as well, use local_install_dev_deps().

See also