Install all dependencies of a package tree

Installs all dependencies of a package tree (or source package file), without installing the package tree itself. It installs the development dependencies as well, specified in the Suggests field of DESCRIPTION.

local_install_dev_deps(
  root = ".",
  lib = .libPaths()[1],
  upgrade = FALSE,
  ask = interactive()
)

Arguments

root
Path to the package tree.

lib
Package library to install the packages to. Note that all dependent packages will the be installed here, even if they are already installed in another library.

upgrade
When FALSE, the default, does the minimum amount of work to give you the latest version of pkg. It will only upgrade packages if pkg or one of its explicitly requires a higher version than what you currently have.

When upgrade = TRUE, will do ensure that you have the latest version of pkg and all its dependencies.

ask
Whether to ask for confirmation when installing a different version of a package that is already installed. Installations that only add new packages never require confirmation.

See also