Query the system requirements for a dev package (and its dependencies)

Returns a character vector of commands to run that will install system requirements for the queried operating system.

local_system_requirements(
  os = NULL,
  os_release = NULL,
  root = ".",
  execute = FALSE,
  sudo = execute,
  echo = FALSE
)

pkg_system_requirements(
  package,
  os = NULL,
  os_release = NULL,
  execute = FALSE,
  sudo = execute,
  echo = FALSE
)

Arguments

os, os_release
The operating system and operating system release version, e.g. "ubuntu", "debian", "centos", "redhat". See https://github.com/rstudio/r-system-requirements#operating-systems for all full list of supported operating systems.

If NULL, the default, these will be looked up using distro::distro().

root
Path to the package tree.

execute, sudo
If execute is TRUE, pak will execute the system commands (if any). If sudo is TRUE, pak will prepend the commands with sudo.

echo
If echo is TRUE and execute is TRUE, echo the command output.

package
The package name to lookup system requirements for.

Value

A character vector of commands needed to install the system requirements for the package.

Examples

if (FALSE) {
pkg_system_requirements("pak", "ubuntu", "20.04")
pkg_system_requirements("pak", "redhat", "7")
}