Install packages from CRAN, Bioconductor, GitHub, URLs, etc.

Install packages from CRAN, Bioconductor, GitHub, URLs, etc. Learn how to tell pak which packages to install, and where those packages can be found.

If you want a quick overview of package sources, see "Get started with pak".

Details

include_docs("pkgdepends", "docs/pkg-refs.rds", top = FALSE)