Draw the dependency tree of a package

Draw the dependency tree of a package

Usage

pkg_deps_tree(pkg, upgrade = TRUE, dependencies = NA)

Arguments

pkg

Package names or package references. E.g.

ggplot2: package from CRAN, Bioconductor or a CRAN-like repository in general,

tidyverse/ggplot2: package from GitHub,

tidyverse/ggplot2@v3.4.0: package from GitHub tag or branch,

https://examples.com/.../ggplot2_3.3.6.tar.gz: package from URL,

.: package in the current working directory.

See "Package sources" for more details.

upgrade

Whether to use the most recent available package versions.

dependencies

What kinds of dependencies to install. Most commonly one of the following values:

NA: only required (hard) dependencies,

TRUE: required dependencies plus optional and development dependencies,

FALSE: do not install any dependencies. (You might end up with a non-working package, and/or the installation might fail.) See Package dependency types for other possible values and more information about package dependencies.

Value

The same data frame as pkg_deps(), invisibly.

Examples

pkg_deps_tree("dplyr")

#> dplyr 1.0.10 âœ¨                                                          
#> â”œâ”€generics 0.1.3 âœ¨                                                      
#> â”œâ”€glue 1.6.2 âœ¨                                                          
#> â”œâ”€lifecycle 1.0.3 âœ¨                                                     
#> â”‚ â”œâ”€cli 3.4.1 âœ¨                                                         
#> â”‚ â”œâ”€glue                                                                
#> â”‚ â””â”€rlang 1.0.6 âœ¨                                                       
#> â”œâ”€magrittr 2.0.3 âœ¨                                                      
#> â”œâ”€R6 2.5.1 âœ¨                                                            
#> â”œâ”€rlang                                                                 
#> â”œâ”€tibble 3.1.8 âœ¨                                                        
#> â”‚ â”œâ”€fansi 1.0.3 âœ¨                                                       
#> â”‚ â”œâ”€lifecycle                                                           
#> â”‚ â”œâ”€magrittr                                                            
#> â”‚ â”œâ”€pillar 1.8.1 âœ¨                                                      
#> â”‚ â”‚ â”œâ”€cli                                                               
#> â”‚ â”‚ â”œâ”€fansi                                                             
#> â”‚ â”‚ â”œâ”€glue                                                              
#> â”‚ â”‚ â”œâ”€lifecycle                                                         
#> â”‚ â”‚ â”œâ”€rlang                                                             
#> â”‚ â”‚ â”œâ”€utf8 1.2.2 âœ¨                                                      
#> â”‚ â”‚ â””â”€vctrs 0.5.1 âœ¨                                                     
#> â”‚ â”‚   â”œâ”€cli                                                             
#> â”‚ â”‚   â”œâ”€glue                                                            
#> â”‚ â”‚   â”œâ”€lifecycle                                                       
#> â”‚ â”‚   â””â”€rlang                                                           
#> â”‚ â”œâ”€pkgconfig 2.0.3 âœ¨                                                   
#> â”‚ â”œâ”€rlang                                                               
#> â”‚ â””â”€vctrs                                                               
#> â”œâ”€tidyselect 1.2.0 âœ¨                                                    
#> â”‚ â”œâ”€cli                                                                 
#> â”‚ â”œâ”€glue                                                                
#> â”‚ â”œâ”€lifecycle                                                           
#> â”‚ â”œâ”€rlang                                                               
#> â”‚ â”œâ”€vctrs                                                               
#> â”‚ â””â”€withr 2.5.0 âœ¨                                                       
#> â”œâ”€vctrs                                                                 
#> â””â”€pillar                                                                
#>                                                                         
#> Key:  âœ¨ new                                                             

pkg_deps_tree("r-lib/usethis")

#> r-lib/usethis 2.1.6.9000 âœ¨ğŸ‘·ğŸ¼ğŸ”                                           
#> â”œâ”€cli 3.4.1 âœ¨                                                           
#> â”œâ”€clipr 0.8.0 âœ¨                                                         
#> â”œâ”€crayon 1.5.2 âœ¨                                                        
#> â”œâ”€curl 4.3.3 âœ¨                                                          
#> â”œâ”€desc 1.4.2 âœ¨                                                          
#> â”‚ â”œâ”€cli                                                                 
#> â”‚ â”œâ”€R6 2.5.1 âœ¨                                                          
#> â”‚ â””â”€rprojroot 2.0.3 âœ¨                                                   
#> â”œâ”€fs 1.5.2 âœ¨                                                            
#> â”œâ”€gert 1.9.2 âœ¨ â¬‡ (1.91 MB)                                              
#> â”‚ â”œâ”€askpass 1.1 âœ¨                                                       
#> â”‚ â”‚ â””â”€sys 3.4.1 âœ¨                                                       
#> â”‚ â”œâ”€credentials 1.3.2 âœ¨ â¬‡ (170.15 kB)                                   
#> â”‚ â”‚ â”œâ”€openssl 2.0.5 âœ¨                                                   
#> â”‚ â”‚ â”‚ â””â”€askpass                                                         
#> â”‚ â”‚ â”œâ”€sys                                                               
#> â”‚ â”‚ â”œâ”€curl                                                              
#> â”‚ â”‚ â”œâ”€jsonlite 1.8.4 âœ¨                                                  
#> â”‚ â”‚ â””â”€askpass                                                           
#> â”‚ â”œâ”€openssl                                                             
#> â”‚ â”œâ”€rstudioapi 0.14 âœ¨                                                   
#> â”‚ â”œâ”€sys                                                                 
#> â”‚ â””â”€zip 2.2.2 âœ¨                                                         
#> â”œâ”€gh 1.3.1 âœ¨ â¬‡ (95.20 kB)                                               
#> â”‚ â”œâ”€cli                                                                 
#> â”‚ â”œâ”€gitcreds 0.1.2 âœ¨ â¬‡ (95.59 kB)                                       
#> â”‚ â”œâ”€httr 1.4.4 âœ¨                                                        
#> â”‚ â”‚ â”œâ”€curl                                                              
#> â”‚ â”‚ â”œâ”€jsonlite                                                          
#> â”‚ â”‚ â”œâ”€mime 0.12 âœ¨                                                       
#> â”‚ â”‚ â”œâ”€openssl                                                           
#> â”‚ â”‚ â””â”€R6                                                                
#> â”‚ â”œâ”€ini 0.3.1 âœ¨ â¬‡ (13.13 kB)                                            
#> â”‚ â””â”€jsonlite                                                            
#> â”œâ”€glue 1.6.2 âœ¨                                                          
#> â”œâ”€jsonlite                                                              
#> â”œâ”€lifecycle 1.0.3 âœ¨                                                     
#> â”‚ â”œâ”€cli                                                                 
#> â”‚ â”œâ”€glue                                                                
#> â”‚ â””â”€rlang 1.0.6 âœ¨                                                       
#> â”œâ”€purrr 0.3.5 âœ¨                                                         
#> â”‚ â”œâ”€magrittr 2.0.3 âœ¨                                                    
#> â”‚ â””â”€rlang                                                               
#> â”œâ”€rappdirs 0.3.3 âœ¨                                                      
#> â”œâ”€rlang                                                                 
#> â”œâ”€rprojroot                                                             
#> â”œâ”€rstudioapi                                                            
#> â”œâ”€whisker 0.4.1 âœ¨ â¬‡ (65.36 kB)                                          
#> â”œâ”€withr 2.5.0 âœ¨                                                         
#> â””â”€yaml 2.3.6 âœ¨                                                          
#>                                                                         
#> Key:  âœ¨ new |  â¬‡ download | ğŸ‘·ğŸ¼ build | ğŸ” compile                        

See also