List all R versions supported by Posit Package Manager (PPM)

List all R versions supported by Posit Package Manager (PPM)

Usage

ppm_r_versions()

Value

Data frame with columns:

r_version: minor R versions, i.e. version numbers containing the first two components of R versions supported by this PPM instance.

See also

Examples

if (FALSE) {
ppm_r_versions()
}