List all available Posit Package Manager (PPM) snapshots

List all available Posit Package Manager (PPM) snapshots

Usage

ppm_snapshots()

Value

Data frame with two columns:

date: the time the snapshot was taken, a POSIXct vector,

id: integer id of the snapshot, this can be used in the repository URL.

Details

The repository URL of a snapshot has the following form on Windows:

{base}/{repo}/{id}

where {base} is the base URL for PPM (see ppm_repo_url()) and {id} is either the date or id of the snapshot, or latest for the latest snapshot. E.g. these are equivalent:

https://packagemanager.posit.co/cran/5
https://packagemanager.posit.co/cran/2017-10-10

On a Linux distribution that has PPM support, the repository URL that contains the binary packages looks like this:

{base}/{repo}/__linux__/{binary_url}/{id}

where {id} is as before, and {binary_url} is a code name for a release of a supported Linux distribution. See the binary_url column of the result of ppm_platforms() for these code names.

See also

Examples

ppm_snapshots()
#> # A data frame: 896 × 2
#>    date                   id
#>    <dttm>              <int>
#>  1 2017-10-10 00:00:00     5
#>  2 2017-10-25 00:00:00     6
#>  3 2017-10-30 00:00:00     7
#>  4 2017-11-01 00:00:00     8
#>  5 2017-11-03 00:00:00     9
#>  6 2017-11-07 00:00:00    10
#>  7 2017-11-08 00:00:00    11
#>  8 2017-11-13 00:00:00    12
#>  9 2017-11-15 00:00:00    13
#> 10 2017-11-16 00:00:00    14
#> # ℹ 886 more rows