Query the currently configured CRAN-like repositories

pak uses the repos option, see options(). It also automatically adds a CRAN mirror if none is set up, and the correct version of the Bioconductor repositories. See the cran_mirror and bioc arguments.

Usage

repo_get(r_version = getRversion(), bioc = TRUE, cran_mirror = NULL)

Arguments

r_version

R version to use to determine the correct Bioconductor version, if bioc = TRUE.

bioc

Whether to automatically add the Bioconductor repositories to the result.

cran_mirror

CRAN mirror to use. Leave it at NULL to use the mirror in getOption("repos") or an automatically selected one.

Details

repo_get() returns the table of the currently configured repositories.

Examples

repo_get()

#> # A data frame: 5 × 5                                                   
#>   name          url                                type  r_ver…¹ bioc_…²
#> * <chr>         <chr>                              <chr> <chr>   <chr>  
#> 1 CRAN          https://cloud.r-project.org        cran  *       NA     
#> 2 BioCsoft      https://bioconductor.org/packages… bioc  4.2.2   3.16   
#> 3 BioCann       https://bioconductor.org/packages… bioc  4.2.2   3.16   
#> 4 BioCexp       https://bioconductor.org/packages… bioc  4.2.2   3.16   
#> 5 BioCworkflows https://bioconductor.org/packages… bioc  4.2.2   3.16   
#> # … with abbreviated variable names ¹​r_version, ²​bioc_version         

See also