Query the currently configured CRAN-like repositories

pak uses the repos option, see options(). It also automatically adds a CRAN mirror if none is set up, and the correct version of the Bioconductor repositores. See the cran_mirror and bioc arguments.

repo_get(r_version = getRversion(), bioc = TRUE, cran_mirror = NULL)

Arguments

r_version
R version to use to determine the correct Bioconductor version, if bioc = TRUE.

bioc
Whether to automatically add the Bioconductor repositories to the result.

cran_mirror
CRAN mirror to use. Leave it at NULL to use the mirror in getOption("repos") or an automatically selected one.

Details

repo_get() returns the table of the currently configured repositories.

See also

Examples

repo_get()

#> # A tibble: 5 x 5
#>   name        url                                   type  r_version bioc_version
#>   <chr>       <chr>                                 <chr> <chr>     <chr>       
#> 1 CRAN        https://cloud.r-project.org           cran  *         NA          
#> 2 BioCsoft    https://bioconductor.org/packages/3.… bioc  4.0.5     3.12        
#> 3 BioCann     https://bioconductor.org/packages/3.… bioc  4.0.5     3.12        
#> 4 BioCexp     https://bioconductor.org/packages/3.… bioc  4.0.5     3.12        
#> 5 BioCworkfl… https://bioconductor.org/packages/3.… bioc  4.0.5     3.12