Show the status of CRAN-like repositories

It checks the status of the configured or supplied repositories.

repo_status(
  platforms = NULL,
  r_version = getRversion(),
  bioc = TRUE,
  cran_mirror = NULL
)

repo_ping(
  platforms = NULL,
  r_version = getRversion(),
  bioc = TRUE,
  cran_mirror = NULL
)

Arguments

platforms

Platforms to use, default is the current platform, plus source packages.

r_version

R version(s) to use, the default is the current R version, via getRversion().

bioc

Whether to add the Bioconductor repositories. If you already configured them via options(repos), then you can set this to FALSE.

cran_mirror

The CRAN mirror to use.

Value

A data frame that has a row for every repository, on every queried platform and R version. It has these columns:

name: the name of the repository. This comes from the names of the configured repositories in options("repos"), or added by pkgcache. It is typically CRAN for CRAN, and the current Bioconductor repositories are BioCsoft, BioCann, BioCexp, BioCworkflows.

url: base URL of the repository.

bioc_version: Bioconductor version, or NA for non-Bioconductor repositories.

platform: platform, possible values are source, macos and windows currently.

path: the path to the packages within the base URL, for a given platform and R version.

r_version: R version, one of the specified R versions.

ok: Logical flag, whether the repository contains a metadata file for the given platform and R version.

ping: HTTP response time of the repository in seconds. If the ok column is FALSE, then this columns in NA.

error: the error object if the HTTP query failed for this repository, platform and R version.

Details

repo_ping() is similar to repo_status() but also prints a short summary of the data, and it returns its result invisibly.

See also

Examples

if (FALSE) {
repo_status()
repo_status(
  platforms = c("windows", "macos"),
  r_version = c("4.0", "4.1")
)
repo_ping()
}