Show the status of CRAN-like repositories

It checks the status of the configured or supplied repositories.

Usage

repo_status(
  platforms = NULL,
  r_version = getRversion(),
  bioc = TRUE,
  cran_mirror = NULL
)

repo_ping(
  platforms = NULL,
  r_version = getRversion(),
  bioc = TRUE,
  cran_mirror = NULL
)

Arguments

platforms

Platforms to use, default is the current platform, plus source packages.

r_version

R version(s) to use, the default is the current R version, via getRversion().

bioc

Whether to add the Bioconductor repositories. If you already configured them via options(repos), then you can set this to FALSE.

cran_mirror

The CRAN mirror to use.

Value

A data frame that has a row for every repository, on every queried platform and R version. It has these columns:

name: the name of the repository. This comes from the names of the configured repositories in options("repos"), or added by pkgcache. It is typically CRAN for CRAN, and the current Bioconductor repositories are BioCsoft, BioCann, BioCexp, BioCworkflows.

url: base URL of the repository.

bioc_version: Bioconductor version, or NA for non-Bioconductor repositories.

platform: platform, possible values are source, macos and windows currently.

path: the path to the packages within the base URL, for a given platform and R version.

r_version: R version, one of the specified R versions.

ok: Logical flag, whether the repository contains a metadata file for the given platform and R version.

ping: HTTP response time of the repository in seconds. If the ok column is FALSE, then this columns in NA.

error: the error object if the HTTP query failed for this repository, platform and R version.

Details

repo_ping() is similar to repo_status() but also prints a short summary of the data, and it returns its result invisibly.

Examples

repo_status()

#> # A data frame: 10 × 10                                                 
#>    name     url   type  bioc_…¹ platf…² path  r_ver…³ ok     ping error 
#>    <chr>    <chr> <chr> <chr>   <chr>   <chr> <chr>   <lgl> <dbl> <list>
#>  1 CRAN     http… cran  NA      source  src/… 4.2     TRUE  0.269 <NULL>
#>  2 CRAN     http… cran  NA      aarch6… bin/… 4.2     TRUE  0.265 <NULL>
#>  3 BioCsoft http… bioc  3.16    source  src/… 4.2     TRUE  0.264 <NULL>
#>  4 BioCsoft http… bioc  3.16    aarch6… bin/… 4.2     TRUE  0.271 <NULL>
#>  5 BioCann  http… bioc  3.16    source  src/… 4.2     TRUE  0.383 <NULL>
#>  6 BioCann  http… bioc  3.16    aarch6… bin/… 4.2     TRUE  0.392 <NULL>
#>  7 BioCexp  http… bioc  3.16    source  src/… 4.2     TRUE  0.507 <NULL>
#>  8 BioCexp  http… bioc  3.16    aarch6… bin/… 4.2     TRUE  0.739 <NULL>
#>  9 BioCwor… http… bioc  3.16    source  src/… 4.2     TRUE  0.505 <NULL>
#> 10 BioCwor… http… bioc  3.16    aarch6… bin/… 4.2     TRUE  0.766 <NULL>
#> # … with abbreviated variable names ¹​bioc_version, ²​platform,         
#> #   ³​r_version                                                         

repo_status(
  platforms = c("windows", "macos"),
  r_version = c("4.0", "4.1")
)

#> # A data frame: 15 × 10                                                 
#>    name          url    type  bioc_…¹ platf…² r_ver…³ path  ok      ping
#>    <chr>         <chr>  <chr> <chr>   <chr>   <chr>   <chr> <lgl>  <dbl>
#>  1 CRAN          https… cran  NA      i386+x… 4.0     bin/… TRUE   0.198
#>  2 CRAN          https… cran  NA      i386+x… 4.1     bin/… TRUE   0.197
#>  3 CRAN          https… cran  NA      aarch6… 4.1     bin/… TRUE   0.197
#>  4 BioCsoft      https… bioc  3.12    i386+x… 4.0     bin/… TRUE   0.988
#>  5 BioCann       https… bioc  3.12    i386+x… 4.0     bin/… TRUE   1.03 
#>  6 BioCexp       https… bioc  3.12    i386+x… 4.0     bin/… TRUE   1.25 
#>  7 BioCworkflows https… bioc  3.12    i386+x… 4.0     bin/… TRUE   1.47 
#>  8 BioCsoft      https… bioc  3.14    i386+x… 4.1     bin/… TRUE   1.48 
#>  9 BioCsoft      https… bioc  3.14    aarch6… 4.1     bin/… FALSE NA    
#> 10 BioCann       https… bioc  3.14    i386+x… 4.1     bin/… TRUE   1.45 
#> 11 BioCann       https… bioc  3.14    aarch6… 4.1     bin/… FALSE NA    
#> 12 BioCexp       https… bioc  3.14    i386+x… 4.1     bin/… TRUE   1.72 
#> 13 BioCexp       https… bioc  3.14    aarch6… 4.1     bin/… FALSE NA    
#> 14 BioCworkflows https… bioc  3.14    i386+x… 4.1     bin/… TRUE   1.71 
#> 15 BioCworkflows https… bioc  3.14    aarch6… 4.1     bin/… FALSE NA    
#> # … with 1 more variable: error <list>, and abbreviated variable names  
#> #   ¹​bioc_version, ²​platform, ³​r_version                             

repo_ping()

#> Repository summary:                   source aarch64-apple-darwin20          
#> CRAN          @ cloud.r-project.org     ✔              ✔              (194ms)
#> BioCsoft      @ bioconductor.org        ✔              ✔              (352ms)
#> BioCann       @ bioconductor.org        ✔              ✔              (511ms)
#> BioCexp       @ bioconductor.org        ✔              ✔              (675ms)
#> BioCworkflows @ bioconductor.org        ✔              ✔              (698ms)

See also