List contents of the system requirements DB, for a platform

It also tries to update the system dependency database, if it is outdated. (I.e. older than allowed in the metadata_update_after configuration option .

Usage

sysreqs_db_list(sysreqs_platform = NULL)

Arguments

sysreqs_platform

System requirements platform. If NULL, then the sysreqs_platform configuration option is used, which defaults to the current platform. Set this option if pak does not detect your platform correctly.

Value

Data frame with columns:

name: cross platform system dependency name in the database.

patterns: one or more regular expressions to match to SystemRequirements fields.

packages: one or more system package names to install.

pre_install: command(s) to run before installing the packages.

post_install:: command(s) to run after installing the packages.

See also

Examples

sysreqs_db_list(sysreqs_platform = "ubuntu-22.04")
#> # A data frame: 125 × 5
#>    name       patterns  packages  pre_install post_install
#>    <chr>      <list>    <list>    <list>      <list>      
#>  1 apparmor   <chr [2]> <chr [1]> <NULL>      <NULL>      
#>  2 atk        <chr [1]> <chr [1]> <NULL>      <NULL>      
#>  3 automake   <chr [1]> <chr [1]> <NULL>      <NULL>      
#>  4 berkeleydb <chr [2]> <chr [1]> <NULL>      <NULL>      
#>  5 blender    <chr [1]> <chr [1]> <NULL>      <NULL>      
#>  6 boost      <chr [1]> <chr [1]> <NULL>      <NULL>      
#>  7 bowtie2    <chr [1]> <chr [1]> <NULL>      <NULL>      
#>  8 bwidget    <chr [1]> <chr [1]> <NULL>      <NULL>      
#>  9 cairo      <chr [1]> <chr [1]> <NULL>      <NULL>      
#> 10 chrome     <chr [2]> <chr [1]> <chr [3]>   <NULL>      
#> # ℹ 115 more rows