List platforms with system requirements support

List platforms with system requirements support

Usage

sysreqs_platforms()

Value

Data frame with columns:

name: human readable OS name.

os: OS name, e.g. linux.

distribution: OS id, e.g. ubuntu or redhat.

version: distribution version. A star means that all versions are supported, that are also supported by the vendor.

update_command: command to run to update the system package metadata.

install_command: command to run to install packages.

query_command: name of the tool to use to query system package information.

See also

Examples

sysreqs_platforms()
#> # A data frame: 10 × 7
#>    name          os    distribution version update_command install_command
#>    <chr>         <chr> <chr>        <chr>   <chr>          <chr>          
#>  1 Ubuntu Linux  linux ubuntu       *       apt-get -y up… apt-get -y ins…
#>  2 Debian Linux  linux debian       *       apt-get -y up… apt-get -y ins…
#>  3 CentOS Linux  linux centos       *       NA             yum install -y 
#>  4 Rocky Linux   linux rockylinux   *       NA             dnf install -y 
#>  5 Red Hat Ente… linux redhat       6       NA             yum install -y 
#>  6 Red Hat Ente… linux redhat       7       NA             yum install -y 
#>  7 Red Hat Ente… linux redhat       *       NA             dnf install -y 
#>  8 Fedora Linux  linux fedora       *       NA             dnf install -y 
#>  9 openSUSE Lin… linux opensuse     *       NA             zypper --non-i…
#> 10 SUSE Linux E… linux sle          *       NA             zypper --non-i…
#> # ℹ 1 more variable: query_command <chr>